Feat: 비회원 기본 재료 추가

+++ src/main/java/com/eateum/eateumbe/fridges/service/FridgeServiceImpl.java
@@ -2,7 +2,6 @@ package com.eateum.eateumbe.fridges.service;
 
 import java.util.List;
 
-import com.eateum.eateumbe.global.common.PageResponse;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
@@ -12,6 +11,7 @@ import com.eateum.eateumbe.fridges.dto.request.FridgeRequest;
 import com.eateum.eateumbe.fridges.dto.response.FridgeResponse;
 import com.eateum.eateumbe.fridges.dto.response.FridgeResponse.AddItem;
 import com.eateum.eateumbe.fridges.repository.FridgeMapper;
+import com.eateum.eateumbe.global.common.PageResponse;
 import org.springframework.web.multipart.MultipartFile;
 
 
@@ -25,6 +25,17 @@ public class FridgeServiceImpl implements FridgeService {
     //내 냉장고 재료 조회
     @Override
     public PageResponse<FridgeResponse> getMyFridgeItems(String userId, int page, int size) {
+
+        //비회원인 경우: DB 조회 없이 고정된 7가지 재료 반환
+        if("guest".equalsIgnoreCase(userId)){
+            List<String> guestItemNames = List.of("돼지고기", "스팸", "달걀", "김치", "라면", "양파", "파");
+
+            List<FridgeResponse> guestList = fridgeMapper.selectItemsByNames(guestItemNames);
+
+            return PageResponse.of(guestList, guestList.size(), page, size);
+        }
+
+        //회원인 경우
         //offset을 설정하기 위함 페이지가 넘어가면 그 앞에 개수를 제외하고 순서대로 재료를 가지고 온다. (무한 스크롤 진행이라도 몇 개인지 기준이 필요함)
         int offset = (page - 1) * size;
         //Mapper에게 데이터 조회를 요청시킨다. (limit, offset을 전달)
@@ -45,6 +56,12 @@ public class FridgeServiceImpl implements FridgeService {
     @Override
     @Transactional
     public AddItem addItem(String userId, FridgeRequest request) {
+
+        //비회원은 재료 추가 불가
+        if("guest".equalsIgnoreCase(userId)){
+            throw new IllegalArgumentException("재료를 저장하려면 로그인이 필요합니다.");
+        }
+
         fridgeMapper.addFridgeItem(userId, request.getItemId());
 
         return fridgeMapper.selectItemDetail(request.getItemId());
@@ -54,6 +71,11 @@ public class FridgeServiceImpl implements FridgeService {
     @Override
     @Transactional
     public void deleteItem(String userId, Long itemId) {
+        //비회원은 재료 추가 불가
+        if("guest".equalsIgnoreCase(userId)){
+            throw new IllegalArgumentException("재료를 삭제하려면 로그인이 필요합니다.");
+        }
+
         fridgeMapper.deleteItem(userId, itemId);
     }
 
@@ -72,7 +94,12 @@ public class FridgeServiceImpl implements FridgeService {
     @Override
     @Transactional
     public void addItems(String userId, List<Long> itemIds) {
-        if(itemIds != null || !itemIds.isEmpty()){
+        //비회원은 재료 추가 불가
+        if("guest".equalsIgnoreCase(userId)){
+            throw new IllegalArgumentException("재료를 저장하려면 로그인이 필요합니다.");
+        }
+
+        if(itemIds != null && !itemIds.isEmpty()){
             fridgeMapper.addFridgeItems(userId, itemIds);
         }
     }
